federation-api: Move create_join_event::v1 test as integration test
Otherwise cargo check always complains that we are using
a deprecated const.
See https://github.com/rust-lang/rust/issues/47238
