+++ crates/ruma-federation-api/tests/membership/mod.rs
@@ -1,0 +1,1 @@
+mod create_join_event;
