+++ crates/ruma-federation-api/tests/tests.rs
@@ -1,0 +1,1 @@
+mod membership;
